pyproject.toml
[project]
name = "crag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
   "langchain>=1.0.0",
   "langchain-community>=0.4.1",
   "langchain-google-genai>=3.0.1",
   "langchain-pinecone>=0.2.13",
   "langchain-tavily>=0.2.13",
   "langchainhub>=0.1.21",
   "langgraph>=1.0.1",
   "langsmith>=0.4.41",
   "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
